feat: Add new code using clean data and generate .gitignore and analisis/requisitos.md with some comments of the results
