+++ analisis/requisitos.md
@@ -1,0 +1,43 @@
+# Análisis de Requisitos - YouTube GB
+
+## Requisito 1-3: Análisis de Categorías y Engagement
+
+* La música domina YouTube GB con **45%** de videos en tendencia y **914M** likes
+* El entretenimiento es segundo pero con menos de la mitad de engagement que música
+* Ratio likes/dislikes muestra que el contenido es generalmente bien recibido en todas las categorías
+
+## Requisito 4: Ratio Vistas/Comentarios
+
+* Los datos muestran una correlación fuerte entre vistas y comentarios
+* La tasa de comentarios es más alta en contenido de entretenimiento y música
+* El engagement mediante comentarios es un indicador fuerte del éxito del video
+
+## Requisito 5: Análisis Temporal
+
+* Los datos muestran patrones claros de actividad por temporada
+* La música mantiene consistencia temporal mientras otras categorías fluctúan
+* Los fines de semana muestran picos de actividad en todas las categorías
+
+## Requisito 6: Análisis de Canales
+
+* Dominan los shows de TV tradicional (Fallon, Kimmel, Ellen)
+* Los top canales mantienen consistencia con ~205-208 videos cada uno
+* Sugiere una estrategia efectiva de migración de contenido TV a YouTube
+
+## Requisito 7: Análisis Geográfico
+
+* La distribución de vistas es uniforme en áreas urbanas principales
+* Londres y alrededores muestran la mayor concentración de actividad
+* El engagement rate es consistente (3.6-4.2%) en todas las regiones
+
+## Requisito 8: Análisis de Comentarios
+
+* Alta correlación entre número de comentarios y éxito del video
+* No es posible análisis de sentimiento por falta de contenido de comentarios
+* Los videos más comentados tienden a tener mejor engagement general
+
+## Requisito 9: Factibilidad de Predicciones
+
+* Baja capacidad predictiva general (R² máximo de 0.41)
+* El conteo de comentarios es el mejor predictor de éxito
+* Se necesitan más variables para mejorar la precisión de predicciones
